feat: move routing form methods to InsightsRoutingBaseService
- Move all RoutingEventsInsights methods to InsightsRoutingBaseService
- Consolidate duplicate objectToCsv functions
- Update trpc-router.ts to use service methods instead of static calls
- Delete routing-events.ts file and associated test
- Clean up unused imports

This refactoring consolidates routing form functionality into the dedicated
service as requested, removing the need for the separate routing-events.ts file.
